Adding RDS Snapshot Auditor.
